WEBAPP-126: Move tests to components/__tests__
