+++ src/modules/layout/components/__tests__/Footer.spec.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Footer from '../components/Footer'
+import Footer from '../Footer'
 import { shallow } from 'enzyme'
 
 describe('Footer', () => {
