+++ src/modules/layout/components/__tests__/GeneralFooter.spec.js
@@ -1,6 +1,6 @@
 import { shallow } from 'enzyme'
 import React from 'react'
-import GeneralFooter from '../components/GeneralFooter'
+import GeneralFooter from '../GeneralFooter'
 
 describe('GeneralFooter', () => {
   test('should match snapshot', () => {
