+++ src/modules/layout/components/__tests__/GeneralHeader.spec.js
@@ -1,6 +1,6 @@
 import { shallow } from 'enzyme'
 import React from 'react'
-import GeneralHeader from '../components/GeneralHeader'
+import GeneralHeader from '../GeneralHeader'
 
 describe('GeneralHeader', () => {
   test('should match snapshot', () => {
